+++ durden/icons/default.lua
@@ -41,7 +41,7 @@ return {
 {
 	[24] = function()
 		return icon_synthesize_src("cli_24px.png", 24,
-			icon_colorize, {color = {"fff", 0.0, 1.0, 1.0}});
+			icon_colorize, {color = {"fff", 1.0, 1.0, 1.0}});
 	end,
 }
 };